+++ src/app/components/transfers/live-stats-table.js
@@ -5,7 +5,7 @@
  *
  * @module components/transfers/live-stats-table
  * @author Michal Borzecki, Jakub Liput
- * @copyright (C) 2017 ACK CYFRONET AGH
+ * @copyright (C) 2017-2018 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -16,8 +16,12 @@ import _ from 'lodash';
 const {
   Component,
   computed,
+  observer,
   inject: {
     service,
+  },
+  run: {
+    scheduleOnce,
   },
   get,
   set,
@@ -82,6 +86,12 @@ export default Component.extend({
   rerunTransfer: undefined,
   
   /**
+   * @virtual 
+   * @type {Function}
+   */
+  notifyTransferListChanged: () => {},
+  
+  /**
    * Type of transfers. May be `scheduled`, `current` or `completed`
    * @public
    * @type {string}
@@ -94,6 +104,11 @@ export default Component.extend({
    * @type {Array<string>|undefined} array of transfer ids
    */
   selectedTransferIds: undefined,
+      
+  /**
+   * @type {EmberObject}
+   */
+  firstRowSpace: undefined,
   
   /**
    * If true, component is rendered in mobile mode.
@@ -142,7 +157,6 @@ export default Component.extend({
    */
   _tableDataCache: null,
   
-  
   _tableData: computed(
     'transfers.[]',
     'transfers.{startIndex,endIndex}',
@@ -174,7 +188,9 @@ export default Component.extend({
       );
       
       if (transfers && providers) {
+        let arrayChanged = false;
         const cachedTransfers = _tableDataCache.mapBy('transfer');
+        const initialCacheLength = get(_tableDataCache, 'length');
         _.pullAllWith(
           _tableDataCache,
           _.difference(
@@ -186,6 +202,10 @@ export default Component.extend({
             return cached && transfer && ct === transfer;
           }
         );
+        if (get(_tableDataCache, 'length') !== initialCacheLength) {
+          arrayChanged = true;
+        }
+        
         transfers.forEach(transfer => {
           if (!_.includes(cachedTransfers, transfer)) {
             const transferId = get(transfer, 'id');
@@ -206,6 +226,7 @@ export default Component.extend({
               updaterId,
               actions: _rowActions,
             }));
+            arrayChanged = true;
           }
         });
         _tableDataCache.sort((a, b) => get(a, 'listIndex') - get(b, 'listIndex'));
@@ -215,7 +236,9 @@ export default Component.extend({
           'firstRowListIndex',
           topRecord && get(topRecord, 'listIndex')
         );
-        _tableDataCache.arrayContentDidChange();
+        if (arrayChanged) {
+          _tableDataCache.arrayContentDidChange();
+        }
       }
       
       return _tableDataCache;
@@ -283,7 +306,7 @@ export default Component.extend({
         id: 'totalFiles',
         propertyName: 'totalFiles',
         title: i18n.t(I18N_PREFIX + 'totalFiles'),
-        component: 'transfers/live-stats-table/cell-errorable',
+        component: 'transfers/live-stats-table/cell-total-files',
       },
       {
         id: 'type',
@@ -415,11 +438,16 @@ export default Component.extend({
     }
     return actions;
   }),
-  
-  /**
-   * @type {EmberObject}
-   */
-  firstRowSpace: undefined,
+    
+  transferListChanged: observer('_tableDataCache.[]', function() {
+    scheduleOnce(
+      'afterRender',
+      this,
+      'notifyTransferListChanged',
+      this.get('transferType')
+    );
+
+  }),
   
   init() {
     this._super(...arguments);